+++ src/services/RevertedBlockResolver.ts
@@ -2,6 +2,7 @@ import {getModelForClass} from '@typegoose/typegoose';
 import {inject, injectable} from 'inversify';
 import newrelic from 'newrelic';
 import Block from '../models/Block';
+import {BlockRevertedEvent} from '../types/ReportedMetricsEvents';
 import {Logger} from 'winston';
 
 @injectable()
@@ -14,7 +15,7 @@ class RevertedBlockResolver {
     }
 
     this.logger.warn(`Block reverted: from ${lastSubmittedBlockId} --> ${nextBlockId}`);
-    newrelic.recordCustomEvent('BlockReverted', {
+    newrelic.recordCustomEvent(BlockRevertedEvent, {
       lastSubmittedBlockId: lastSubmittedBlockId,
       nextBlockId: nextBlockId,
     });